[enhancement/10] Rename index.ts → decode-json.ts
